MCOL-4255 1.5 use static_cast for int to uint

+++ utils/funcexp/func_round.cpp
@@ -140,7 +140,7 @@ uint64_t Func_round::getUintVal(Row& row,
     if (UNLIKELY(op_ct.colDataType == execplan::CalpontSystemCatalog::DATE))
     {
         IDB_Decimal d = getDecimalVal(row, parm, isNull, op_ct);
-        x = d.value;
+        x = static_cast<uint64_t>(d.value);
     }
     else
     {
